refactor(cli-run): split event handlers by concern
